Using JQuery to modify and create new HTML Elements
JQuery is a versatile language that is essentially JavaScript, except more concise. This program is the code from an extremely basic game I created that is similar to Cookie Clicker or Make It Rain. There is no story line, but your main objective is to make the page as long as possible by clicking on a button. Every time you click a button, a paragraph is inserted into a div, making the page longer. There is a click counter to display number of clicks, as well as a reset button. Feel free to edit the code!